+++ src/packages/media/media/components/index.ts
@@ -2,3 +2,4 @@ import './input-media/index.js';
 
 export * from './input-media/index.js';
 export * from './input-image-cropper/index.js';
+export * from './input-upload-field/index.js';
