+++ src/packages/media/media/components/input-upload-field/input-upload-field-image.element.ts
