+++ src/packages/media/media/components/input-upload-field/input-upload-field-svg.element.ts
