+++ src/packages/media/media/components/input-upload-field/input-upload-field.element.ts
@@ -1,6 +1,6 @@
-import type { MediaValueType } from '../../../property-editors/upload-field/property-editor-ui-upload-field.element.js';
-import type { UmbTemporaryFileModel } from '../../temporary-file/temporary-file-manager.class.js';
-import { TemporaryFileStatus, UmbTemporaryFileManager } from '../../temporary-file/temporary-file-manager.class.js';
+import type { MediaValueType } from '../../property-editors/upload-field/property-editor-ui-upload-field.element.js';
+import { TemporaryFileStatus, UmbTemporaryFileManager } from '@umbraco-cms/backoffice/temporary-file';
+import type { UmbTemporaryFileModel } from '@umbraco-cms/backoffice/temporary-file';
 import { UmbId } from '@umbraco-cms/backoffice/id';
 import {
 	css,
